refactor(power): introduce an internal `reset` module
